+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [0.9.15] - UNRELEASED
 
+- Added help buttons on every dialog
 - Refined warning and error messages to be more clear and reflect the [docs](https://github.com/hakasapl/PGPatcher/wiki/Error-Message-Guide)
 
 ## [0.9.14] - 2026-02-21